# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        example: Description of the error
        type: string
    type: object
  handlers.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/handlers.PaginationDetails'
    type: object
  handlers.PaginationDetails:
    properties:
      limit:
        example: 20
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  models.UserBlueprintProgress:
    properties:
      consumed:
        type: boolean
      created_at:
        type: string
      id:
        type: integer
      item:
        $ref: '#/definitions/models.Item'
      item_id:
        type: integer
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relations
      user_id:
        type: integer
    type: object
  models.UserHideoutModuleProgress:
    properties:
      created_at:
        type: string
      hideout_module:
        $ref: '#/definitions/models.HideoutModule'
      hideout_module_id:
        type: integer
      id:
        type: integer
      level:
        type: integer
      unlocked:
        type: boolean
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relations
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Server for Arc Project (Mobile/Web)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ArcAPI
  version: "1.0"
paths:
  /admin/export/alerts:
    get:
      description: Fetch all alert data in CSV format. Only admins can export data.
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file content
          schema:
            type: string
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Export alerts
      tags:
      - management
  /admin/export/bots:
    get:
      description: Fetch all bot data in CSV format. Only admins can export data.
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file content
          schema:
            type: string
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Export bots
      tags:
      - management
  /admin/export/enemy-types:
    get:
      description: Fetch all enemy type data in CSV format. Only admins can export
        data.
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file content
          schema:
            type: string
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Export enemy types
      tags:
      - management
  /admin/export/hideout-modules:
    get:
      description: Fetch all hideout module data in CSV format. Only admins can export
        data.
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file content
          schema:
            type: string
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Export hideout modules
      tags:
      - management
  /admin/export/items:
    get:
      description: Fetch all item data in CSV format. Only admins can export data.
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file content
          schema:
            type: string
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Export items
      tags:
      - management
  /admin/export/maps:
    get:
      description: Fetch all map data in CSV format. Only admins can export data.
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file content
          schema:
            type: string
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Export maps
      tags:
      - management
  /admin/export/projects:
    get:
      description: Fetch all project data in CSV format. Only admins can export data.
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file content
          schema:
            type: string
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Export projects
      tags:
      - management
  /admin/export/quests:
    get:
      description: Fetch all quest data in CSV format. Only admins can export data.
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file content
          schema:
            type: string
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Export quests
      tags:
      - management
  /admin/export/skill-nodes:
    get:
      description: Fetch all skill node data in CSV format. Only admins can export
        data.
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file content
          schema:
            type: string
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Export skill nodes
      tags:
      - management
  /admin/export/traders:
    get:
      description: Fetch all trader data in CSV format. Only admins can export data.
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file content
          schema:
            type: string
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Export traders
      tags:
      - management
  /admin/sync/force:
    post:
      consumes:
      - application/json
      description: Trigger an immediate synchronization of Game Data from GitHub.
        Only admins can force a sync.
      produces:
      - application/json
      responses:
        "200":
          description: Sync triggered successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Sync already running
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Force data sync
      tags:
      - sync
  /admin/sync/status:
    get:
      consumes:
      - application/json
      description: Check if a synchronization process is currently running. Only admins
        can check status.
      produces:
      - application/json
      responses:
        "200":
          description: Sync status
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get sync status
      tags:
      - sync
  /admin/users/{id}/progress:
    get:
      consumes:
      - application/json
      description: Fetch a comprehensive summary of all progress types (quests, modules,
        skills, blueprints) for any user. Only admins can access this.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched comprehensive progress
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get all user progress
      tags:
      - management
  /admin/users/{id}/progress/blueprints:
    get:
      consumes:
      - application/json
      description: Fetch blueprint consumption states for any user. Only admins can
        access this.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched blueprint progress
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get user blueprint progress
      tags:
      - management
  /admin/users/{id}/progress/blueprints/{item_id}:
    put:
      consumes:
      - application/json
      description: Update blueprint consumption state for any user. Only admins can
        perform this.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item External ID (Blueprint)
        in: path
        name: item_id
        required: true
        type: string
      - description: Consumption status
        in: body
        name: consumption
        required: true
        schema:
          additionalProperties:
            type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated user blueprint progress
          schema:
            $ref: '#/definitions/models.UserBlueprintProgress'
        "400":
          description: Invalid input or ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: User or Blueprint not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update user blueprint progress
      tags:
      - management
  /admin/users/{id}/progress/hideout-modules:
    get:
      consumes:
      - application/json
      description: Fetch hideout module levels for any user. Only admins can access
        this.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched hideout module progress
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get user hideout module progress
      tags:
      - management
  /admin/users/{id}/progress/hideout-modules/{module_id}:
    put:
      consumes:
      - application/json
      description: Update module level/unlock state for any user. Only admins can
        perform this.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Module External ID
        in: path
        name: module_id
        required: true
        type: string
      - description: Progress data
        in: body
        name: progress
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated user module progress
          schema:
            $ref: '#/definitions/models.UserHideoutModuleProgress'
        "400":
          description: Invalid input or ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Not an administrator
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: User or Module not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update user hideout module progress
      tags:
      - management
...
  /alerts/active:
    get:
      consumes:
      - application/json
      description: Fetch all alerts that are currently marked as active
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched active alerts
          schema:
            allOf:
            - $ref: '#/definitions/handlers.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Alert'
                  type: array
              type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List active alerts
      tags:
      - alerts
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"